rename and expand
